Flesh out step componentization.
