+++ app/components/application/application.pug
@@ -8,7 +8,7 @@ main(data-component="application" data-flow="column" data-active-step="embedCode
   .steps(data-ref="stepsContainer")
     include ./steps/intro.pug
     include ./steps/embed-code.pug
-    include ./steps/attributes.pug
+    .step(data-ref="attributePickerMount")
     include ./steps/preview.pug
     include ./steps/details.pug
     include ./steps/creating.pug
