+++ app/components/application/attribute-picker/attribute-picker.pug
@@ -1,10 +1,10 @@
-.step(data-ref="steps[]" data-step="attributes")
+section.step(data-ref="steps[]" data-step="attributePicker")
   form(data-ref="attributesForm")
     h1.headline Choose your options.
 
     p Click parts of your embed code to turn them into options for the user.
 
-    .attribute-picker.box(data-ref="attributePicker")
+    .picker.box(data-ref="picker")
 
     .attribute-list-mount(data-ref="attributeListMount")
