+++ app/components/application/attribute-picker/attribute-picker.styl
@@ -1,0 +1,88 @@
+@import "~styl/colors"
+@import "~styl/fonts"
+
+.step[data-step="attributePicker"]
+  .picker
+    color mutedTextColor
+    cursor default
+    overflow-wrap break-word
+    user-select none
+    white-space pre-wrap
+    word-break break-word
+    word-wrap break-word
+    // Fix animation artifacts.
+    background #fff
+    subpixelAntialiasedFonts()
+
+    [data-entity-id]
+      display inline-block
+      position relative
+      border-radius 2em
+      padding 0 .5em
+      line-height 1.3
+      font-family inherit
+      cursor pointer
+      box-shadow 0 0 0 1px rgba(mutedTextColor, .5)
+      background #eee
+
+      &:focus
+        outline none
+        box-shadow 0 0 2px 1px saturation(brandOrangeDark, 75%)
+
+      &:not(.tracked)
+        &:hover
+          box-shadow 0 0 0 1px rgba(textColor, .8)
+
+      &.tracked
+        background #fff
+        box-shadow 0 0 0 2px brandOrange
+
+        &:focus
+          box-shadow 0 0 2px 2px brandOrangeDark
+
+    .hljs-string, .hljs-number
+      color textColor
+
+    .details
+      color darken(mutedTextColor, 20%)
+      font-family sansSerifFonts
+      font-size 1.2em
+      margin 0
+      text-align center
+      white-space normal
+      padding-left 6.5em
+      padding-right @padding-left
+
+      + .details
+        margin-top -2.8em
+
+    [data-chunk-type="param-group"]
+      [data-chunk-type="param-key"]
+        color darkGray
+
+      &.tracked [data-chunk-type="param-key"]
+        color brandOrange
+
+    [data-chunk-type="entity-group"].tracked
+      .hljs-attr, .entity-identifier
+        color brandOrange
+
+    [data-chunk-type="attribute-group"].tracked
+      .hljs-attr
+        color brandOrange
+
+  p.divider
+    padding-top 2em
+
+  .embed-code-location-container
+    display block
+
+  .embed-code-location
+    padding 1.25em
+    width 100%
+
+    .input
+      margin 0
+
+      & + .input
+        margin-top 1.3125em
