--- app/components/application/steps/attributes.styl
@@ -1,16 +1,0 @@
-.step[data-step="attributes"]
-  p.divider
-    padding-top 2em
-
-  .embed-code-location-container
-    display block
-
-  .embed-code-location
-    padding 1.25em
-    width 100%
-
-    .input
-      margin 0
-
-      & + .input
-        margin-top 1.3125em
