+++ app/components/application/steps/embed-code.pug
@@ -32,4 +32,4 @@
 
   .actions
     button.button.secondary.transparent(data-ref="navigationButtons[]" data-step="intro") Back
-    button.button.primary.next(data-ref="navigationButtons[]" data-step="attributes") Next
+    button.button.primary.next(data-ref="navigationButtons[]" data-step="attributePicker") Next
