+++ app/components/application/steps/preview.pug
@@ -6,5 +6,5 @@
     .preview-container(data-ref="previewContainer")
 
   .actions
-    button.button.secondary.transparent(data-ref="navigationButtons[]" data-step="attributes") Back
+    button.button.secondary.transparent(data-ref="navigationButtons[]" data-step="attributePicker") Back
     button.button.primary.next(data-ref="navigationButtons[]" data-step="details") Next
